Feature/security headers (#5)
* add security headers
via https://developer.okta.com/blog/2019/04/11/site-security-cloudflare-netlify#netlify-to-the-rescue

* add security headers
- always check the console first, that'd speed things up 🤦🏾‍♂️

*ToDo*
Missing Headers
Permissions Policy is a new header that allows a site to control which features and APIs can be used in the browser.

Warnings
This policy contains 'unsafe-inline' which is dangerous in the default-src directive. This policy contains 'unsafe-inline' which is dangerous in the style-src directive.